Updated the MVC tutorial for React 16
I had issues that many others have had with getting the server rendering of this tutorial to work out of the box. I came to realize that it was a discrepancy in the version of React being used on the client versus in the NuGet package. NuGet was using the latest, version 16, but the client side code all referenced version 15. All sorts of confusion and chaos ensued.

This is a major change but it boils down to a shift over to ES6 classes instead of the older `React.createClass()` method. I tested as I went along, so the code should all work. I did not update screenshots. I think any reasonably seasoned .NET developer will skip over those details anyway. I am not a seasoned .NET developer, and I had no issues with that side of things.

I also added a note at the bottom explaining how server-side rendering can suffer from this version drift issue. Hopefully that will save people hours of debugging and frustration.