Avoid crash if setTimeout/clearTimeout is referenced
In some cases, React (or other libraries) will assume that setTimeout
and clearTimeout are defined on the global scope, without invoking it.

We should still throw an error in case a consumer expects this to just
work on the server (it won't), but we can avoid a crash by defining the
function on the global scope.

For context: https://github.com/reactjs/React.NET/issues/555