Fix exception handling during render in .NET Core
If an exception is thrown outside of the render call in
Razor views (async invocation), the app could crash. This does not appear
to be an issue in ASP.NET.

Unfortunately this means extra string allocations to hold the render
result, but the rest of the StringBuilder logic used internally is not
a problem.