Bump template versions automatially during release
This is a bit of a chicken-and-egg problem. Because this is a monorepo, the templates live alongside the source. Therefore when releasing a new version, the template needs to be validated against the not-yet-published version of React.AspNet.

May need to adjust the strategy as more templates are added, but this at least improves the release story so updates to the core libraries do not require two releases for the updates to make it into the published templates.